Simplify bindgen.patch.
Every time a new allowlist-item is added to the bindgen_flags list, we have to rebase. I reworked this part of the patch to remove any reference to them and to just make our own bindgen_flags. The original bindgen_flags get put in a not_needed variable so the build doesn't fail