various gcc build fixes - most pulled from upstream
A bunch of simple header include fixes, with the exception
of the return of fixes/absl-optional.patch. This is an older patch
which works around some issues with clang using gcc's <optional>
header. The old version in this repo was using gcc 13's optional
header; this was updated for gcc 14.